bricklink_telegram_bot/requirements.txt
requests
rauth
python-telegram-bot
boto3